Revert the old chat events back for fix chat placeholder replacing
Fixes #1118

+++ src/main/java/com/gamingmesh/jobs/commands/list/leave.java
@@ -41,7 +41,7 @@ public class leave implements Cmd {
 		Util.LEAVECONFIRM.add(uuid);
 
 		plugin.getServer().getScheduler().runTaskLater(plugin, () -> Util.LEAVECONFIRM.remove(uuid),
-		    (long) (20 * Jobs.getGCManager().ConfirmExpiryTime));
+		    20 * Jobs.getGCManager().ConfirmExpiryTime);
 
 		pSender.sendMessage(Jobs.getLanguage().getMessage("command.leave.confirmationNeed", "[jobname]",
 		    job.getNameWithColor(), "[time]", Jobs.getGCManager().ConfirmExpiryTime));
